Move the liveness updater and singleton deployer into core
The liveness machinery consumes only core primitives and now maintains any
ServiceDirectory implementation through the new updateLiveness and
reconcileLiveness API methods. The deployer deploys nothing when no
ServiceDirectory bean is present, preserving the standalone-core
does-nothing invariant.
